LayoutTests/applicationmanifest/developer-warnings.html

<link rel="manifest" href="resources/developer-warnings.manifest">
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
    testRunner.getApplicationManifestThen(function() {
        alert("Pass");
        testRunner.notifyDone();
    });
}
</script>
