LayoutTests/applicationmanifest/developer-warnings-expected.txt

CONSOLE MESSAGE: Parsing application manifest developer-warnings.manifest: The value of "start_url" is not a valid URL.
CONSOLE MESSAGE: Parsing application manifest developer-warnings.manifest: The value of "name" is not a string.
ALERT: Pass
